+++ dservercore/utils_auth.py
@@ -11,6 +11,7 @@ from flask import current_app
 
 from flask_jwt_extended import jwt_required as flask_jwt_required
 from flask_jwt_extended import get_jwt_identity as flask_get_jwt_identity
+from flask_jwt_extended import get_jwt as flask_get_jwt
 
 
 def jwt_required(*jwt_required_args, **jwt_required_kwargs):
@@ -32,6 +33,14 @@ def get_jwt_identity():
         return current_app.config.get("DEFAULT_USER")
     else:
         return flask_get_jwt_identity()
+
+
+def get_jwt():
+    """Return JWT payload or empty dict if JWT authorisation disabled."""
+    if current_app.config.get("DISABLE_JWT_AUTHORISATION"):
+        return {}
+    else:
+        return flask_get_jwt()
 
 
 def _get_user_obj(username):